fix(order-service): corrige erros de compilação e validação de valor monetário
- Ajusta comparação de BigDecimal usando compareTo ao invés de operadores (<=)
- Padroniza uso de BigDecimal no campo amount
- Corrige inconsistência entre DTO e Service
- corrigido status CANCELLED ao enum OrderStatus
- Melhora validação de valor garantindo maior que zero

+++ order-service/src/main/java/com/jonathanleite/vitrine/orderservice/dto/OrderRequestDTO.java
@@ -13,7 +13,7 @@ public class OrderRequestDTO {
 
     @NotNull(message = "Valor é obrigatório")
     @DecimalMin(value = "0.01", message = "O valor deve ser maior que zero")
-    private Double amount;
+    private BigDecimal amount;
 
     public OrderRequestDTO() {
     }
